refactor(update): isolate externalization bridge handoff
